Added the ability to mass change up to 6 NPC stats at once by name, race, class, bodytype, HP and Level or any combination thereof. Or, every NPC in the zone can be changed at once.
Added the same but specifically for special_abilities with the option of adding or removing a single ability, or applying the raw database column of the current NPC to any NPC matched the parameters.

Both of the new options can be accessed by selecting a NPC, and clicking the gear icon on the top left of the page.

Fixed some minor issues and added a missing column to NPCs.

Removed out of era entries in some defines.