Support for new wp_spawns spawngroup feature; font enlargement
Spawngroups have a new option: a 'WP Spawns' boolean that when enabled will make every spawnpoint with a grid in that spawngroup spawn the NPCs on a random waypoint taken from their grids (if they have a grid) and ignore the spawnpoint coordinates for that spawnpoint.  This is intended to be used for grids set to Random (or similar types) and should particuarly be used for fast respawning outdoor NPCs.

I also increased the font size for the 'edit_form_content' class selector, (and the header) which will make all kinds of boxes a little bigger.  I made another selector to retain the old size and applied it to the NPC edit pages so they would remain the same size.  Frankly I think the text is a bit too small as this project was started before 1080p was common, but if this change makes some forms look bad, the smaller selector can be used or this change can be easily reverted.