Greed field added to NPC types; Add NPC form fixes
NPC types now have a greed field which is for merchant greed.  Default value is 0.  Max value is 255 and most greedy merchants will be somewhere between 5-100.  Most merchants should have this at 0.

The Add New NPC form was fixed and default values were improved to better match a typical NPC of the entered level.  It's still missing some fields though. (like avoidance)

I added some isset() checks and initialized values to certain variables to prevent some Notice warnings when developer errors are on for the NPC stats pages.

If attempting to view stats on an NPC ID that doesn't exist, the table header will now indicate that the NPC does not exist.