New feature: importing grids from delimited text data
The view grid page now has a text box that allows you to paste in coordinate data (or simply edit the grid in text instead of web forms) which will replace the current grid when submitted.  The coordinate data is comma, tab, or space delimited.

Why is this awesome?  Because now you can compile a coordinate list in a text editor and paste it in to create grids which makes replicating Sony's grids MUCH easier/faster.  I have also created a lua script to parse myseq timer files so I can sort and format coordinates for pasting.

The procedure is: log in players with mercenaries on aggressive and park them in places where they will kill the most NPCs in a roambox or roaming path.  Over many hours myseq will then figure out where all the grid points are (this takes a few days).  You can then open the timer file, copy and paste it and figure out which coordinates are in the roambox and which aren't. (again I made a lua script tool to do this for me which excludes spawns outside a specified box and/or specified NPCs)  Then you format the text to paste into this editor and viola, your grid matches Sony's